# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ledgerline API
  version: 1.0.0
  description: Multi-tenant double-entry accounting API.
servers:
  - url: http://localhost:3000/api/v1
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            requestId:
              type: string
          required: [code, message]
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      required: [id, email, createdAt]
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        role:
          type: object
          description: Current user's role in this organization. Present on list responses.
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required: [id, name]
        permissions:
          type: array
          description: Current membership's permission codes. Present on list responses.
          items:
            type: string
      required: [id, name, isActive, createdAt]
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        role:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        isActive:
          type: boolean
      required: [id, user, role, isActive]
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        isControlAccount:
          type: boolean
        isBankAccount:
          type: boolean
        isActive:
          type: boolean
      required: [id, code, name, type, isControlAccount, isBankAccount, isActive]
    TaxCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        rate:
          type: string
          description: Exact decimal rate serialized as a string.
          example: "0.1300"
        type:
          type: string
          enum: [VAT, EXEMPT, ZERO]
        isActive:
          type: boolean
      required: [id, code, name, rate, type, isActive]
    Party:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [customer, supplier, both]
        code:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        creditDays:
          type: integer
        isActive:
          type: boolean
      required: [id, type, code, name, isActive]
    FiscalYear:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isClosed:
          type: boolean
      required: [id, label, startDate, endDate, isClosed]
    Period:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fiscalYearId:
          type: string
          format: uuid
        label:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isOpen:
          type: boolean
      required: [id, fiscalYearId, label, startDate, endDate, isOpen]
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    OrganizationHeader:
      name: X-Organization-Id
      in: header
      required: true
      description: Which organization this request operates on. Verified server-side against an active membership.
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /orgs:
    get:
      summary: List organizations the current user belongs to
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
    post:
      summary: Create a new organization (creator becomes Owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
  /orgs/{id}/members:
    get:
      summary: List members of an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Add a member to the organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                roleId:
                  type: string
                  format: uuid
              required: [email, roleId]
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '403':
          $ref: '#/components/responses/Forbidden'
  /accounts:
    get:
      summary: List the organization's chart of accounts
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      summary: Create an account
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                name:
                  type: string
                type:
                  type: string
                  enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
              required: [code, name, type]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
  /parties:
    get:
      summary: List customers/suppliers, with search and pagination
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: List of parties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Party'
    post:
      summary: Create a customer or supplier
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [customer, supplier, both]
                code:
                  type: string
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                creditDays:
                  type: integer
              required: [type, code, name]
      responses:
        '201':
          description: Party created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '400':
          $ref: '#/components/responses/BadRequest'
  /tax-codes:
    get:
      summary: List active tax codes for invoice entry
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      responses:
        '200':
          description: Active tax codes ordered by code
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxCode'
  /parties/{id}:
    patch:
      summary: Update a customer or supplier
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: false
              properties:
                type:
                  type: string
                  enum: [customer, supplier, both]
                code:
                  type: string
                  minLength: 1
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                  nullable: true
                phone:
                  type: string
                  nullable: true
                creditDays:
                  type: integer
                  minimum: 0
                isActive:
                  type: boolean
      responses:
        '200':
          description: Party updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /fiscal-years:
    get:
      summary: List fiscal years
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      responses:
        '200':
          description: List of fiscal years
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FiscalYear'
  /periods:
    get:
      summary: List accounting periods
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: fiscalYearId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of accounting periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Period'
...